avicii-common: prop: Disable vsync for CPU rendered Apps
* Choreographer VSync has some performance impact for CPU rendered apps
as the app has to wait for VSync to start a new frame draw. When the
frame draw time exceeds one VSync, this has a direct impact on the FPS.
debug.cpurend.vsync property controls the VSync for CPU rendered apps.

Disable VSync for CPU rendered app, by setting this property to false.

Signed-off-by: chandu078 <chandudyavanapelli03@gmail.com>
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>
Change-Id: I26a149cfa75052ab08d0ac11196a89f87461e201